+++ CHANGELOG.md
@@ -22,6 +22,7 @@ New URI functions and an updated dependency for finding logos that may be a brea
 * Two example scripts have been included in the repo. They can be used to demonstrate and test logo detection.
 * Metadata with no Organization now returns an Organization struct that has a name and displayname 
   of "Unknown" in English. Previously name and displayname maps were empty.
+* If a ROR URL/ID is already present in an Organization record it is downloaded directly, without searching
 
 ### Other Changes
 * Some tests have been temporarily disabled or changed to account for CERN search results from ROR changing